// lib/database/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
/*
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */

export const LeadStatus = {
  LEAD_FRAIS: "LEAD_FRAIS",
  A_RAPPELER: "A_RAPPELER",
  PAS_INTERESSE: "PAS_INTERESSE",
  EN_ATTENTE: "EN_ATTENTE",
  RDV: "RDV",
  NRP_1: "NRP_1",
  NRP_2: "NRP_2",
  DEVIS_A_CREER: "DEVIS_A_CREER",
  DEVIS_ENVOYE: "DEVIS_ENVOYE",
  CONTRAT_SIGNE: "CONTRAT_SIGNE",
} as const;

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadPriority = {
  FAIBLE: "FAIBLE",
  MOYENNE: "MOYENNE",
  HAUTE: "HAUTE",
} as const;

export type LeadPriority = (typeof LeadPriority)[keyof typeof LeadPriority];